Thomas/template/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Sean Sica <23294618+seansica@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
fastapi = {extras = ["standard"], version = "^0.112.2"}
joblib = "1.4.2"
scikit-learn = "1.5.2"
fastapi-cache2 = {version = "0.1.9", extras = ["redis"]}

[tool.poetry.group.dev.dependencies]
uvicorn = "^0.30.6"
pytest = "^8.3.2"
ruff = "^0.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
